Fix evented_socket test for connection timeout
This patch parameterizes the host and port in the evented_socket tests and
updates the test for connection timeouts to use a non-routable IP address
instead of a server socket that has not called listen. This fixes a test
failure on Linux where connections to bound ports are refused instead of
stalled if listen has not been called.